Fixed tests for actions
