+++ test/mocha.opts
@@ -1,2 +1,3 @@
+--recursive
 --require test/babeloptions
 --grep babeloptions --invert
